+++ index.html
@@ -89,8 +89,8 @@
           <label class="ad-form__label" for="title">Заголовок объявления</label>
           <input id="title" name="title" type="text" placeholder="Милая, уютная квартирка в центре Токио" required minlength="30" maxlength="100"
           data-pristine-required-message="Обязательное поле"
-          data-pristine-minlength-message="Минимальная длина 2 символа"
-          data-pristine-maxlength-message="Максимальная длина 25 символов">
+          data-pristine-minlength-message="Минимальная длина 30 символов"
+          data-pristine-maxlength-message="Максимальная длина 100 символов">
         </fieldset>
         <fieldset class="ad-form__element ad-form__element--wide">
           <label class="ad-form__label" for="address">Адрес (координаты)</label>
@@ -108,7 +108,9 @@
         </fieldset>
         <fieldset class="ad-form__element">
           <label class="ad-form__label" for="price">Цена за ночь, руб.</label>
-          <input id="price" name="price" type="number" placeholder="5000" required max="100000">
+          <input id="price" name="price" type="number" placeholder="5000" required max="100000"
+          data-pristine-required-message="Обязательное поле"
+          data-pristine-max-message="Максимальная цена 100 000 руб.">
           <div class="ad-form__slider"></div>
         </fieldset>
         <fieldset class="ad-form__element ad-form__element--time">
