+++ js/form.js
@@ -1,66 +1,32 @@
-import {isEscapeKey, isEnterKey} from './util.js';
 import { offerNameByType } from './card-popup.js';
 
 const adForm = document.querySelector('.ad-form');
 const adFormElements = adForm.children;
 const mapFilters = document.querySelector('.map__filters');
 const mapFiltersElements = mapFilters.children;
-const mainBody  = document.querySelector('body');
+
 
 // Блокировка полей форм
-const toggleDisabledElements = (arrayOfElements) => {
-  for (const element of arrayOfElements) {
+
+/**
+ * Проверяет на блокировку элементы формы.
+ * @param {array} Elements Элементы формы.
+ */
+function toggleDisabledElements(Elements) {
+  for (const element of Elements) {
     element.disabled = !element.disabled;
   }
-};
+}
 
-function toggleReadOnly (){
+/**
+ * Переключает активность состояния формы.
+ */
+export function toggleReadOnly (){
   adForm.classList.toggle('ad-form--disabled');
   mapFilters.classList.toggle('map__filters--disabled');
   toggleDisabledElements(adFormElements);
   toggleDisabledElements(mapFiltersElements);
 }
-
-
-// Сообщения об ошибке/успехе валидации
-const FormSuccess = 'success';
-const FormError = 'error';
-
-const removeValidationMessage = (message) => {
-  const messageElement = document.querySelector(`.${message}`);
-  mainBody.removeChild(messageElement);
-};
-
-const createValidationMessage = (message) => {
-  const messageTemplate = document.querySelector(`#${message}`)
-    .content
-    .querySelector(`.${message}`);
-  const messageElement = messageTemplate.cloneNode(true);
-
-  if (message === FormError) {
-    const messageErrorButton = messageElement.querySelector('.error__button');
-    messageErrorButton.addEventListener('click', () => {
-      removeValidationMessage(message);
-    });
-    messageErrorButton.addEventListener('keydown', (evt) => {
-      if (isEnterKey(evt)) {
-        removeValidationMessage(message);
-      }
-    });
-  }
-  messageElement.addEventListener('click', () => {
-    removeValidationMessage(message);
-  });
-  document.addEventListener('keydown', (evt) => {
-    if (isEscapeKey(evt)) {
-      evt.preventDefault();
-      removeValidationMessage(message);
-    }
-  });
-  mainBody.appendChild(messageElement);
-};
-
-createValidationMessage(FormError);
 
 // Валидация формы
 const pristine = new Pristine(adForm, {
@@ -69,12 +35,6 @@ const pristine = new Pristine(adForm, {
   errorTextClass: 'ad-form__error-text',
   errorTextTag: 'span',
 });
-
-function validateTittle (value) {
-  return value.length >= 30 && value.length <= 100;
-}
-
-pristine.addValidator(adForm.querySelector('#title'), validateTittle,'От 2 до 50 символов');
 
 const priceField = adForm.querySelector('#price');
 const typeField = adForm.querySelector('[name="type"]');
@@ -92,7 +52,7 @@ function validatePrice () {
 
 function getPriceErrorMessage () {
   return `
-  Для типа жилья "${offerNameByType[typeField.value]}" минимальная цена - ${minPriceDictionary[typeField.value]} руб.
+  Для типа жилья "${offerNameByType[typeField.value]}" минимальная цена - ${minPriceDictionary[typeField.value].toLocaleString('ru')} руб.
   `;
 }
 pristine.addValidator(priceField, validatePrice, getPriceErrorMessage);
@@ -109,4 +69,3 @@ adForm.addEventListener('submit', (evt) => {
   pristine.validate();
 });
 
-export {toggleReadOnly};
